==> 04_summary_report_senate.Rmd <==
---
title: "2020 Election Early Voting Breakdowns - Key Senate Races"
author: "Aaron Kessler, Investigative Data Team"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(janitor)
library(lubridate)
library(readxl)
library(writexl)
library(tidycensus)
library(kableExtra)
options(scipen = 999)
options(dplyr.summarise.inform=F)


#run processing script to refresh with latest data as needed each day
# source("01_load_data.R")
# source("02_analysis.R")

# alternatively, use saved versions of files already created (much faster when new updated data not needed)
state_bothcycles <- readRDS("processed_data/state_bothcycles.rds")
# county_bothcycles <- readRDS("processed_data/county_bothcycles.rds")
natl_grandtots_bothyears <- readRDS("processed_data/natl_grandtots_bothyears.rds")
state_grandtots_bothyears <- readRDS("processed_data/state_grandtots_bothyears.rds")
# county_grandtots_bothyears <- readRDS("processed_data/county_grandtots_bothyears.rds")

# load lookup file with key state designations for prez and senate
keystates_lookup <- read_excel("processed_data/keystates_elex2020.xlsx")


#pull national votes cast/return number for use below
value_national_2016 <- natl_grandtots_bothyears %>% pull(x2016)
value_national_2020 <- natl_grandtots_bothyears %>% pull(x2020)
value_national_pctchg <- natl_grandtots_bothyears %>% pull(pctchg_returned) %>% round_half_up(1)

#we'll use use prez battlegrounds, create a vector to filer on
keystates_prez_vector <- keystates_lookup %>% 
  filter(office == "P") %>% 
  distinct(state) %>% 
  pull()

#use those values to create a subset of the states table
state_keyprezonly_bothcycles <- state_bothcycles %>% 
  filter(state %in% keystates_prez_vector)

state_keyprezonly_bothcycles


#we'll use use senate key races, create a vector to filer on
keystates_senate_vector <- keystates_lookup %>% 
  filter(office == "S") %>% 
  distinct(state) %>% 
  pull()

#use those values to create a subset of the states table
state_keysenateonly_bothcycles <- state_bothcycles %>% 
  filter(state %in% keystates_senate_vector)

state_keysenateonly_bothcycles



#### CHOOSE WHAT FILTERED DATASET TO USE FOR ANALYSIS
#select which filtered version of the full dataset to use below, or create new one
votedata <- state_keysenateonly_bothcycles


#note: something strange happening with PA. Will troubleshoot and until then just taking it out for now
votedata <- votedata %>% 
  filter(state != "PA")



```

Bloomberg has an agreement with Catalist, a data company that works with Democrats and others, to compile counts of individual ballots cast before Election Day, either early in-person or by-mail, across 40 U.S. states.  
  
The data is compiled from individual-level voting records in those states and is being provided to us each weekday morning. (Because of when and how certain states report their aggregate totals vs. individual voter-level information, you may see slight variations between Catalist's data and the AP's daily statewide tallies).  
  
Our agreement allows us to use this data for stories and other news content across all platforms, provided we simply cite Catalist as the data source. If there are questions about what you see below, or you have ideas for other kinds of more targeted breakdowns that would help you in your stories or interactives, please drop a line at akessler39@bloomberg.net.

Below are breakdowns of the **ballots cast as of the morning of `r format(Sys.Date(), "%B %d, %Y")`**, in states with this year's most competitive Senate races... 

Bottom line: there are huge surges among voters across the board in battleground states, particularly among young voters and minority voters. Meanwhile, in states where party affiliation is identified, Democrats continue to be well ahead of Republicans in their growth from 2016 in most of those states.
  
## Statewide Totals

```{r, echo=FALSE}

#note: something strange happening with PA. Will troubleshoot and until then just taking it out for now
state_grandtots_bothyears <- state_grandtots_bothyears %>% 
  filter(state != "PA")


state_grandtots_bothyears %>% 
  filter(state %in% keystates_prez_vector) %>% 
  mutate(pctchg_returned = round_half_up(pctchg_returned, 1)) %>% 
  select(state,
    "ballots cast at this point 2016" = total_returned_2016,
    "ballots cast to date 2020" = total_returned_2020,
    difference = diff_returned,
    "percent change" = pctchg_returned) %>% 
  kable(format.args = list(big.mark = ",")) %>% #formatting adds (,) 1000s separator
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left") %>% 
  column_spec(1, bold = T)


```

###
## Party Affiliation

```{r, echo=FALSE}

# votedata

party_states <- votedata %>% 
  filter(!state %in% c("TX", "GA")) %>%  #take out these states since no party ID there
  group_by(cycle, state, party_affiliation) %>% 
  summarise(
    total_returned = sum(ballots_returned, na.rm = TRUE)
  ) %>% 
  pivot_wider(names_from = c(cycle), values_from = total_returned) %>% 
  clean_names() %>% 
  mutate(
    diff = x2020 - x2016,
    pctchg = round_half_up((x2020 - x2016) / x2016 * 100, 1)
  ) 

# party_states
  
```


Growth in early voters by their party affiliation to date, compared with the same point before the 2016 election. (Note that some states, such as Texas and Georgia, do not report this information associated with their early voting.)

```{r, echo=FALSE}


# party_states %>% 
#   select(state, party_affiliation, pctchg) %>% 
#   filter(party_affiliation %in% c("Republican", "Democrat")) %>% 
#   mutate(pctchg = round_half_up(pctchg, 1)) %>% 
#   pivot_wider(names_from = party_affiliation, values_from = pctchg) %>%
#   kable(format.args = list(big.mark = ",")) %>% #formatting adds (,) 1000s separator
#   kable_styling(bootstrap_options = c("striped"),
#                 full_width = F,
#                 position = "left") %>%
#   column_spec(1, bold = T)



```


```{r, echo=FALSE}

#for display 
party_states %>% 
  filter(party_affiliation %in% c("Republican", "Democrat")) %>%
  select(
    state,
    party_affiliation,
    "ballots cast at this point in 2016" = x2016,
    "ballots cast so far in 2020" = x2020,
    "difference" = diff,
    "pct change" = pctchg
  ) %>% 
  kable(format.args = list(big.mark = ",")) %>% #formatting adds (,) 1000s separator
  kable_styling(bootstrap_options = c("striped", "condensed"),
                full_width = F,
                position = "left") %>%
  column_spec(1, bold = T)


```

  
  
    
###
## Demographic breakdowns

### Gender

```{r, echo=FALSE}

#calculate growth for each gender group
gender_states <- votedata %>% 
  group_by(cycle, state, gender) %>% 
  summarise(
    total_returned = sum(ballots_returned, na.rm = TRUE)
  ) %>% 
  pivot_wider(names_from = c(cycle, gender), values_from = total_returned) %>% 
  clean_names() %>% 
  mutate(
    diff_female = x2020_female - x2016_female,
    pctchg_female = round_half_up((x2020_female - x2016_female) / x2016_female * 100, 1),
    diff_male = x2020_male - x2016_male,
    pctchg_male = round_half_up((x2020_male - x2016_male) / x2016_male * 100, 1),
    diff_unknown = x2020_unknown - x2016_unknown,
    pctchg_unknown = round_half_up((x2020_unknown - x2016_unknown) / x2016_unknown * 100, 1)
  ) 

#add totals for each cycle and pct totals
gender_states <- gender_states %>% 
  mutate(
    x2016_total = x2016_female + x2016_male + x2016_unknown,
    x2020_total = x2020_female + x2020_male + x2020_unknown,
    x2016_pcttotal_female = round_half_up(x2016_female / x2016_total * 100, 1),
    x2020_pcttotal_female = round_half_up(x2020_female / x2020_total * 100, 1),
    x2016_pcttotal_male = round_half_up(x2016_male / x2016_total * 100, 1),
    x2020_pcttotal_male = round_half_up(x2020_male / x2020_total * 100, 1),
    x2016_pcttotal_unknown = round_half_up(x2016_unknown / x2016_total * 100, 1),
    x2020_pcttotal_unknown = round_half_up(x2020_unknown / x2020_total * 100, 1),
    pct_chg_disparity_male_female = abs(pctchg_female - pctchg_male)
  )

# head(gender_states)

```


Growth in male vs. female early voters 

```{r, echo=FALSE}

#for displaying
gender_states %>% 
  select(
    state,
    "additional female votes cast" = diff_female,
    "additional male votes cast" = diff_male,
    "pct change in female voters" = pctchg_female,
    "pct change in male voters" = pctchg_male
  ) %>% 
  kable(format.args = list(big.mark = ",")) %>% #formatting adds (,) 1000s separator
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left") %>% 
  column_spec(1, bold = T) %>% 
  add_header_above(c(" ", "Differences from 2016" = 2, "Percent Change from 2016" = 2))


```


### Age Groups

```{r, echo=FALSE}

# votedata

age_states <- votedata %>% 
  group_by(cycle, state, age) %>% 
  summarise(
    total_returned = sum(ballots_returned, na.rm = TRUE)
  ) %>% 
  pivot_wider(names_from = c(cycle), values_from = total_returned) %>% 
  clean_names() %>% 
  mutate(
    diff = x2020 - x2016,
    pctchg = round_half_up((x2020 - x2016) / x2016 * 100, 1)
  ) 

# age_states  
  
```


Growth in early voters by age groups to date, compared with the same point before the 2016 election. The table below shows the *percent change* in 2020 compared with 2016 at the same point before the election.

```{r, echo=FALSE}


#for display 
age_states %>% 
  select(state, age, pctchg) %>% 
  filter(age != "Unknown") %>% 
  mutate(pctchg = round_half_up(pctchg, 1)) %>% 
  pivot_wider(names_from = age, values_from = pctchg) %>%
  kable(format.args = list(big.mark = ",")) %>% #formatting adds (,) 1000s separator
  kable_styling(bootstrap_options = c("striped"),
                full_width = F,
                position = "left") %>%
  column_spec(1, bold = T) %>% 
  add_header_above(c(" ", "Percent Change from 2016" = 6))



```

One thing to note is that for the enormous percentage jumps in 18-21 year old voters, they almost all start from a very low 2016 turnout early. Though it's still notable of course just how much that group - traditionally unreliable for voting - has been gangbusters voting early this time.
  
Here below is an expanded table showing the particulars behind those percent changes around ages 18-21.   

```{r, echo=FALSE}

#for display 
age_states %>% 
  filter(age == "18-21") %>% 
  select(
    state,
    age,
    "ballots cast at this point in 2016" = x2016,
    "ballots cast so far in 2020" = x2020,
    "difference" = diff,
    "pct change" = pctchg
  ) %>% 
  kable(format.args = list(big.mark = ",")) %>% #formatting adds (,) 1000s separator
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left") %>% 
  column_spec(1, bold = T)


```


### Race

```{r, echo=FALSE}

race_states <- votedata %>% 
  group_by(cycle, state, race) %>% 
  summarise(
    total_returned = sum(ballots_returned, na.rm = TRUE)
  ) %>% 
  pivot_wider(names_from = c(cycle), values_from = total_returned) %>% 
  clean_names() %>% 
  mutate(
    diff = x2020 - x2016,
    pctchg = round_half_up((x2020 - x2016) / x2016 * 100, 1)
  ) 

# race_states
  
```


Growth in early voters by modeled racial groups to date, compared with the same point before the 2016 election. The table below shows the *percent change* in 2020 compared with 2016 at the same point before the election.

*(Note: this designation is based on a statistical model by Catalist.)*

```{r, echo=FALSE}


#for display 
race_states %>% 
  select(state, race, pctchg) %>% 
  filter(race %in% c("Black", "Caucasian", "Hispanic")) %>% 
  mutate(pctchg = round_half_up(pctchg, 1)) %>% 
  pivot_wider(names_from = race, values_from = pctchg) %>%
  kable(format.args = list(big.mark = ",")) %>% #formatting adds (,) 1000s separator
  kable_styling(bootstrap_options = c("striped"),
                full_width = F,
                position = "left") %>%
  column_spec(1, bold = T) %>% 
  add_header_above(c(" ", "Percent Change from 2016" = 3))



```
  
Below is an expanded table showing more details behind the changes around Black, Hispanic and Caucasion early voters in each state.

```{r, echo=FALSE}

#for display 
race_states %>% 
  filter(race %in% c("Black", "Caucasian", "Hispanic")) %>% 
  select(
    state,
    race,
    "ballots cast at this point in 2016" = x2016,
    "ballots cast so far in 2020" = x2020,
    "difference" = diff,
    "pct change" = pctchg
  ) %>% 
  kable(format.args = list(big.mark = ",")) %>% #formatting adds (,) 1000s separator
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left") %>% 
  column_spec(1, bold = T)


```


```{r, echo=FALSE}




```


```{r, echo=FALSE}




```


```{r, echo=FALSE}




```


```{r, echo=FALSE}




```


```{r, echo=FALSE}




```
